requirements.txt
beautifulsoup4==4.14.2
lxml
panda==0.3.1
psycopg2
numpy==2.2.6
umap-learn==0.5.9.post2
scikit-learn==1.7.2
google-generativeai==0.7.2
fastapi==0.119.0
uvicorn==0.37.0
jinja2==3.1.6
redis==5.0.1
librosa==0.11.0
neucodec>=0.0.4
numpy==2.2.6
phonemizer==3.3.0
soundfile==0.13.1
torch==2.8.0
transformers==4.56.1
resemble-perth==1.0.1
ipykernel==7.0.1
requests==2.32.5
matplotlib-inline==0.1.7
regex==2025.9.18
markdown

